+++ src/java/jp/ac/jec/jz/gr03/entity/Tag.java
@@ -8,6 +8,6 @@ public class Tag {
     public Integer tagId = null;
     public Music music = new Music();
     public String name = null;
-    public Float scoreAverage = null;
+    public Double scoreAverage = null;
     public Integer scoreCount = null;
 }
